Update readme and bump version
I don't have a releases/changelog scheme yet, so I'll leave the
release notes here for now.

This update includes a ~10x performance improvement by saving and
reloading copies of the inventory. I'm sure it can be done faster
yet, perhaps by creating a custom inventory that tracks changes to
itself, but this was the easy way. It's performant enough that the
main bottleneck is the game's code now and not mine. 🎉