Even more performance using recursion and state restoration
Added a secondary algorithm that uses recursion to find solutions.

The original linear algorithm was somewhat slow because it would
break the decision tree into many linear paths that it would explore
one-by-one, meaning that the first choice to either spin the wheel
or purchase a new wheel was performed exactly the same many, many
times.

The recursive algorithm tries to be a bit smarter and ties into
the game much more closely. A new class keeps track of changes to
the inventory as they happen so they can be reverted later.
Another class keeps track of relevant fields for the progress
manager. Together, this means that it can now explore the possible
paths as a tree, rather than linearly. Based entirely on eyeballing
the output speed, this seems to result in about a 4x speed increase.